drishti-mother-child/src/main/java/org/ei/drishti/service/reporting/rules/IsDeathDueToBleedingRule.java
package org.ei.drishti.service.reporting.rules;

import org.ei.drishti.util.SafeMap;
import org.springframework.stereotype.Component;

import static org.ei.drishti.common.AllConstants.ECCloseFields.BLEEDING_VALUE;
import static org.ei.drishti.common.AllConstants.ECCloseFields.MATERNAL_DEATH_CAUSE_FIELD_NAME;

@Component
public class IsDeathDueToBleedingRule implements IRule {

    @Override
    public boolean apply(SafeMap reportFields) {
        return BLEEDING_VALUE.equalsIgnoreCase(reportFields.get(MATERNAL_DEATH_CAUSE_FIELD_NAME));
    }
}

